Add rock-paper-scissors console functionality
gameBoard —> factory function consisting of the board itself in an
array. The IIFE will return the array (formatted) as well as a method
to insert a chosen move onto the board.

Player —> factory function consisting of the player itself, containing
their name and their respective symbol (‘O’ or ‘X’ in this case)

Cell -> factory function which holds a symbol value, and allows you to
modify this value within the function

gameController -> factory function consisting of printing player turns
and playing rounds (game logic to be coded)