Complete console version of tic tac toe
Several pieces of code were implemented, the main one being the
'playGame' function and the 'detectWinner' function.

The 'playGame' function serves as the main exterior of the project,
calling tic tac toe rounds from the 'playRound' function and keeping
track of turns as well. In addition to this, the 'detectWinner' function
was implemented in order to, like it was named, detect the winner of the
tic tac toe game. Inside the function, a 'locations' object is utilized
in order to keep track of the number of similar symbols within a
particular row/column/diagonal. If the number of symbols reached 3
(or -3 for player 2) in any row/column/diagonal, the player who placed
the move is deemed the winner. Rounds are played until one of the
players wins or there are no more moves on the board, which makes the
game a tie.