Complete web + console version of tic tac toe
Finished the web + console hybrid implementation of tic tac toe.

In order to create the board/display for the tic tac toe game, a factory
function was created with functions to display the board as well as
change the interface of the board each time each of the players made a
move.

In addition to this, the gameController FF was updated in order to fit
with the website version of the game. The key change was the
implementation of event listeners to check when each of the players
clicked on a spot on the board. When one of the players did so, the
interface would update accordingly and the program would run as it would
in the console. (check if there are 3 symbols in a row/column/diagonal
using the hashmap/object solution)

A few adjustments were also made throughout the code for simplicity.